requirements.txt
torch
numpy
Pillow
requests